+++ source/commands/utils/_tests/file-utils.test.ts
@@ -1,0 +1,26 @@
+let mockDangerfilePath = ""
+jest.mock("fs", () => { return { existsSync: (p) => p === mockDangerfilePath } })
+
+import { dangerfilePath } from "../file-utils"
+
+describe("dangerfilePath", () => {
+
+  it("should return anything passed into the program's dangerfile", () => {
+    expect(dangerfilePath({ dangerfile: "123"})).toEqual("123")
+  })
+
+  it("should find a dangerfile.js if there is no program, and the .js file exists", () => {
+    mockDangerfilePath = "dangerfile.js"
+    expect(dangerfilePath({})).toEqual("dangerfile.js")
+  })
+
+  it("should find a dangerfile.ts if there is no program, and the .js file does not exist", () => {
+    mockDangerfilePath = "dangerfile.ts"
+    expect(dangerfilePath({})).toEqual("dangerfile.ts")
+  })
+
+  it("should raise if nothing exists", () => {
+    mockDangerfilePath = "dangerfile.tsjs"
+    expect(() => dangerfilePath({}) ).toThrow()
+  })
+})