+++ source/runner/Dangerfile.ts
@@ -2,6 +2,8 @@ import { DangerRuntimeContainer } from "../dsl/DangerResults"
 import { DangerDSLType } from "../dsl/DangerDSL"
 import { MarkdownString } from "../dsl/Aliases"
 
+/// Start of Danger DSL definition
+
 export interface DangerContext {
   /**
    * Contains asynchronous code to be run after the application has booted.
@@ -55,6 +57,8 @@ export interface DangerContext {
   results: DangerRuntimeContainer
 }
 
+/// End of Danger DSL definition
+
 /** Creates a Danger context, this provides all of the global functions
  *  which are available to the Danger eval runtime.
  *
@@ -76,14 +80,38 @@ export function contextForDanger(dsl: DangerDSLType): DangerContext {
   const message = (message: MarkdownString) => results.messages.push({ message })
   const markdown = (message: MarkdownString) => results.markdowns.push(message)
 
-  return {
+  // Anything _but_ danger, that is on the root-level DSL
+  const globals = {
     schedule,
     fail,
     warn,
     message,
     markdown,
     console,
-    results,
-    danger: dsl
+    results
+  }
+
+  // OK, so this is a bit weird, but hear me out.
+  //
+  // I am not sure if it makes sense for "danger js plugins" ( which will
+  // be normal npm modules) to work with the magic globals available in the runtime.
+  //
+  // So I'm _probably_ going to advocate that you pass in the `danger` object into
+  // functions for danger plugins. This means that they can use `danger.fail` etc. This
+  // should make it significantly easier to build and make tests for your modules.
+  //
+  // Which should mean we get more plugins overall.
+  //
+  // Which should be cool.
+  //
+  // So, I'm not going to expose these on the interfaces (and thus the public reference
+  // but it will go into a 'plugin authors guide' whatever that looks like.)
+  //
+  return {
+    ...globals,
+    danger: {
+      ...dsl,
+      ...globals
+    }
   }
 }