+++ source/runner/_tests/fixtures/__DangerfileTypeScript.ts
@@ -1,0 +1,6 @@
+/* tslint-disable */
+
+// This doesn't exist in JS-world
+interface MyThing {}
+
+message("Honey, we got Types")